discord_app/requirements.txt
discord.py==1.3.4
django-environ==0.4.5
feedparser==6.0.1
requests==2.23.0